requirements.txt
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
pandas== 2.0.3
pyodbc==4.0.39
rasa==3.6.4
SQLAlchemy==1.4.49
streamlit== 1.25.0
streamlit-chat== 0.1.1